--- presets/test.js
@@ -1,3 +1,0 @@
-const { encrypt62 } = require("../services/encrypt.service");
-
-console.log(encrypt62(1));
