+++ utils/handlers/jwt.handlers.js
@@ -1,6 +1,6 @@
 const jwt = require("jsonwebtoken");
 
-const logger = require("../logger");
+const logger = require("../../logger");
 const { JWT_SECRET } = require("./config.json").SERVER;
 
 const parseBearerFromHeader = (req, res, next) => {
