+++ utils/handlers/socket.handlers.js
@@ -1,4 +1,4 @@
-const logger = require("../logger");
+const logger = require("../../logger");
 
 const socketErrorHandler = (socket, on, err) => {
   logger.error(
