+++ utils/presets/test.js
@@ -1,0 +1,3 @@
+const { encrypt62 } = require("../encrypt.util");
+
+console.log(encrypt62(1));
